Crate verity_ic

This module serves as the main entry point for the verity_ic library. It includes various submodules that provide functionality related to the Verity project and the Internet Computer ecosystem.

Modules

crypto

This submodule contains cryptographic related operations.

owner

This submodule manages the owner of a canister and provides functionality for authenticated/guarded canister methods.

random

The ‘random’ submodule provides functionality for initializing and retrieving random numbers for use on the Internet Computer (IC).

remittance

The ‘remittance’ submodule contains logic for the remittance canister

verify

whitelist

The ‘whitelist’ module provides CRUD operations for managing a whitelist. The whitelist is a HashMap where each principal is mapped to a boolean value. A principal is considered whitelisted if it maps to true.