verity_ic::crypto

Module ethereum

Functions

get_address_from_public_key

Convert a compressed SEC1 public key (33 bytes) to an Ethereum address (20 bytes)

get_recovery_id

get_signature

Append a recovery identifier byte to the ECDSA signature

hash_eth_message

Preprocess and hash an Ethereum message using the Ethereum message prefix

recover_address_from_eth_signature

Recover the Ethereum address from a given signature and message

sign_message

Sign a message using ECDSA and return the signature