verity_ic

Module remittance

The ‘remittance’ submodule contains logic for the remittance canister

Modules

external_router

state

types

utils

Functions

get_available_balance

Retrieves the available balance for an account associated with a canister

get_canister_balance

Retrieves the canister balance for a specific token and chain

get_reciept

Get the reciept of a successfull withdrawal

get_withheld_balance

Retrieves the withheld balance for an account on a specified canister and chain

init

Initializes the remittance canister state

name

Returns the name of the contract

owner

Returns the owner of the contract

public_key

Get the public key associated with this particular canister

remit

Creates a remittance request

subscribe_to_dc

Subscribes to a DC canister using its canister_id

subscribe_to_pdc

Subscribes to a PCD canister using its canister_id

update_remittance

Validates and updates user and canister balances