verity_ic::verify::types

Type Alias VerificationCanisterResponse

pub type VerificationCanisterResponse = Result<VerificationResponse, String>;

The response from the managed verifier canister. It is a Result type that contains either a VerificationResponse on success or a String error message on failure.

Aliased Type

enum VerificationCanisterResponse {
    Ok(VerificationResponse),
    Err(String),
}

Variants

Ok(VerificationResponse)

Contains the success value

Err(String)

Contains the error value